+++ src/renderer/src/services/ocr/OcrService.ts
